add authentication service

+++ apps/user-service/src/user/user.service.ts
@@ -2,6 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { CreateUserDto, PrismaService, UserDto } from '@app/common';
 import { User } from '@prisma/client';
 import * as bcrypt from 'bcrypt';
+import { UserNotFoundException } from '../exceptions/user-not-found.exception';
+import { InvalidCredantialsException } from '../exceptions/invalid-credentials.exception';
 
 @Injectable()
 export class UserService {
@@ -18,8 +20,30 @@ export class UserService {
     return new UserDto(user);
   }
 
+  async checkCredentials(email: string, password: string) {
+    const user = await this.findOneByEmail(email);
+    if (!user) {
+      throw new InvalidCredantialsException();
+    }
+
+    const hashedPassword = user.password;
+    const isMatch = await bcrypt.compare(password, hashedPassword);
+    if (!isMatch) {
+      throw new InvalidCredantialsException();
+    }
+    return this.prisma.excludeFieldsFromObject(user, [
+      'password',
+      'updatedAt',
+      'createdAt',
+    ]);
+  }
+
   async findOneByEmail(email: string) {
     return await this.prisma.user.findUnique({ where: { email } });
+  }
+
+  async findOneById(id: number) {
+    return await this.prisma.user.findUnique({ where: { id } });
   }
 
   async findAllUsers() {
@@ -28,6 +52,10 @@ export class UserService {
   }
 
   async deleteUser(id: number) {
+    const user = await this.findOneById(id);
+    if (!user) {
+      throw new UserNotFoundException();
+    }
     await this.prisma.user.delete({
       where: { id },
     });
